Type:fix（16579）
Description:When entering the system settings for the first time, there is still a special effect mode in the operation mode, but it can't be selected. It needs to be removed. At present, there is no special effect option in the control panel.
RootCause:Null

Solution:Just a test edition.